code restructuring (#333)
* code restructuring

* comment edits
