+++ src/main/resources/fxml/generate_offline_map_with_local_basemap_dialog.fxml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2019 Esri.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+
+<?import javafx.scene.control.ButtonType?>
+<?import javafx.scene.control.Dialog?>
+<?import javafx.scene.control.DialogPane?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.VBox?>
+<fx:root type="javafx.scene.control.Dialog" xmlns:fx="http://javafx.com/fxml">
+    <dialogPane>
+        <DialogPane>
+            <content>
+                <VBox spacing="6">
+                    <Label text="This web map references a local basemap with name: "/>
+                    <Label fx:id="referencedBasemapFileNameLabel"/>
+                    <Label text="You can locate this basemap on disk or download the basemap again."/>
+                </VBox>
+            </content>
+            <ButtonType fx:id="cancelButton" buttonData="CANCEL_CLOSE" text="Cancel"/>
+            <ButtonType fx:id="localBasemapButton" buttonData="NEXT_FORWARD" text="Choose local basemap"/>
+            <ButtonType fx:id="downloadBasemapButton" buttonData="FINISH" text="Download basemap"/>
+        </DialogPane>
+    </dialogPane>
+</fx:root>
